swap between configurations via "Connection Type"
previously, we always attempted to connect via all configured connection
types, but required that JavaScript connection-type details were filled
out.

This should not be required. Instead, give a "Connection Type" option,
analogous to the "Connection Type" listed in the "Integration" section
of the Referral Candy account area. Only ask for those configuration
options which are relevant to the selection Connection Type.

Then, use the Connection Type selection to determine which method we
attempt to use to talk to Referral Candy.

This change was started when, during testing, it was noticed that the
only way to acquire an email address to forward invoices to, appeared to
be to select the "Email" Connection Type, disabling the "Javascript"
Connection Type, in the "Integration" section of the Referral Candy
account area.